requirements.txt
requests>=2.31.0
websocket-client>=1.6.0
python-telegram-bot>=20.0
pyyaml>=6.0
colorama>=0.4.6
gate-api>=0.4.6
python-dotenv>=1.0.0
